demo/realtime-img2img/requirements.txt
diffusers==0.24.0
transformers==4.35.2
--extra-index-url https://download.pytorch.org/whl/cu121;
torch
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
Pillow==10.1.0
accelerate==0.24.0
compel==2.0.2
controlnet-aux==0.0.7
peft==0.6.0
xformers; sys_platform != 'darwin' or platform_machine != 'arm64'
markdown2
stable_fast @ https://github.com/chengzeyi/stable-fast/releases/download/v0.0.15.post1/stable_fast-0.0.15.post1+torch211cu121-cp310-cp310-manylinux2014_x86_64.whl
